refactor(terminal): tidy up packge
