+++ terminal/imgprotocol/term_detect.go
