+++ terminal/imgprotocol/term_protocol.go
