+++ terminal/imgprotocol/term_protocol_test.go
