+++ terminal/imgprotocol/term_render.go
