+++ terminal/progress/term_progress_writer.go
