+++ terminal/style/term_style_osc4.go
